streamlit-app/requirements.txt
streamlit>=1.28.0
tensorflow-cpu>=2.16.1
torch>=2.0.1
transformers>=4.31.0
pandas>=2.0.3
numpy>=1.24.3
scikit-learn>=1.3.0
nltk>=3.8.1
plotly>=5.15.0
matplotlib>=3.7.2
seaborn>=0.12.2
Pillow>=10.0.0